+++ protected/humhub/modules/installer/views/layouts/main.php
@@ -2,6 +2,7 @@
 
 use yii\helpers\Html;
 use humhub\assets\AppAsset;
+use humhub\widgets\PoweredBy;
 
 /* @var $this \yii\web\View */
 /* @var $content string */
@@ -37,7 +38,7 @@ AppAsset::register($this);
         <?php echo $content; ?>
 
         <div class="text text-center powered">
-            Powered by <a href="http://www.humhub.org" target="_blank">HumHub</a>
+            <?= PoweredBy::widget(); ?>
             <br/>
             <br/>
         </div>