+++ protected/humhub/views/mail/layouts/text.php
@@ -7,4 +7,4 @@
 <?= Yii::t('base', 'Unsubscribe') ?>: <?= $url ?>
 <?php endif; ?>
 
-Powered by HumHub (http://www.humhub.org)
+<?= \humhub\widgets\PoweredBy::widget(['textOnly' => true]); ?>