+++ protected/humhub/widgets/views/footerNavigation_login.php
@@ -6,6 +6,7 @@
  */
 
 use yii\helpers\Html;
+use humhub\widgets\PoweredBy;
 
 $i = 0;
 ?>
@@ -22,5 +23,5 @@ $i = 0;
         <?php endforeach; ?>
         <br/><br/>
     <?php endif; ?>
-    Powered by <?= Html::a('HumHub', 'https://humhub.org'); ?>&nbsp;
+    <?= PoweredBy::widget(); ?>
 </div>