+++ protected/humhub/docs/CHANGELOG.md
@@ -6,6 +6,7 @@ HumHub Change Log
 ---------------------------
 - Enh: Added maximum username length & maximum/minimum space url length (rekollekt)
 - Fix: Error message during database installation
+- Enh: "Powered by" message handling by widget
 
 
 1.3.6  (October 11, 2018)